Split logging and telemetry tests out of protocol_test.exs / plug_test.exs
Neither "protocol error logging" (in protocol_test.exs) nor "telemetry"
(in plug_test.exs) actually test what their host file is meant to cover
- protocol_test.exs is for RFC/wire-protocol conformance and
plug_test.exs is for Plug API compliance, but both of these describe
blocks test cross-cutting Bandit-specific observability features
(configurable log verbosity, :telemetry event emission), using protocol
errors or plug requests merely as a convenient trigger rather than as
the actual subject under test.

Moves them verbatim into two new files:

- logging_test.exs: the 7 tests from protocol_test.exs's "protocol
  error logging" describe block, covering log_protocol_errors and
  log_client_closures verbosity configuration
- telemetry_test.exs: the 13 tests from plug_test.exs's "telemetry"
  describe block, covering :telemetry event emission for start/stop/
  exception events

Each new file is a standalone ExUnit module (since the plug under test
is `__MODULE__` in each file), so local copies of the small plug
helpers each moved describe block depended on (`sleep_and_send`,
`send_200`, etc.) were carried along too - no shared test infrastructure
was refactored, just relocated.

"request isolation" and "process concerns" in plug_test.exs test
similar cross-cutting concerns (process/logger hygiene between
requests, GC behavior, EXIT signal handling) but are left in place, as
is "plug error handling" (whose logging-verbosity assertions are
tightly intertwined with Plug-adapter-contract assertions in the same
tests, making a clean split not mechanical).

Total test count is unchanged (204 across the http1 test directory);
this is purely a file reorganization. Updates lib/bandit/http1/README.md,
which also had a stale reference to a "request_test.exs" that hasn't
existed for some time.
